stable version, next to solve bw for feature

+++ src/hw/hls_target.h
@@ -11,17 +11,17 @@ void hls_target(
 		PackedStencil<dtype, DATAWIDTH, 1, 1, 1>* arg_1,//[34*126*32],
 		PackedStencil<dtype, DATAWIDTH, 1, 1, 1>* arg_2,//[32*32*9]
 		PackedStencil<dtype, DATAWIDTH, 1, 1, 1>* arg_3,//[32*32*9]
-		uint8_t Ksz,
-		uint8_t Xsz,
-		uint8_t Ysz,
-		uint8_t X_n,
-		uint8_t Y_n,
-		uint8_t Cin_n,
-		uint8_t Cin_SZ,
-		uint8_t Cout_n,
-		uint8_t Cout_SZ,
-        uint8_t Stride,
-        uint8_t Ch_Iter,
+		uint16_t Ksz,
+		uint16_t Xsz,
+		uint16_t Ysz,
+		uint16_t X_n,
+		uint16_t Y_n,
+		uint16_t Cin_n,
+		uint16_t Cin_SZ,
+		uint16_t Cout_n,
+		uint16_t Cout_SZ,
+        uint16_t Stride,
+        uint16_t Ch_Iter,
 		bool pool
 );
